Updates for N-API support
It is not possible in N-API to throw a JS exception if you don't have a
`&Context` handle around. So neon-serde's Rust-error-to-JS-exception
conversion unfortunately can't work anymore in N-API.

This patch proposes a different approach: neon-serde's `Result<>` type
gets a new method, `.or_throw(&mut cx)`. With this, users need to
explicitly convert neon-serde `Result`s into neon `JsResult`s.  This
pattern exists in neon core in a few places, like downcasting, so users
should be familiar with it.

The downside to this approach is that folks need to import the
`ThrowExt` trait to get access to the `.or_throw()` method. We might
want to move this trait into the neon prelude later, so that neon-serde
could implement it from there, and users don't need to separately import it.